requirements.txt
# Core FastAPI and web dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# LangChain for AI agent functionality
langchain>=0.0.350
langchain-community>=0.0.38
langchain-openai>=0.0.2
openai>=1.3.0

# HTTP client and API tools
httpx>=0.25.0
aiofiles>=23.0.0

# Data validation and serialization
pydantic>=2.5.0

# OpenAPI client generation
openapi-python-client>=0.15.0

# Logging and monitoring
loguru>=0.7.0
structlog>=23.0.0

# Database and storage (for caching/sessions if needed)
aiosqlite>=0.19.0

# CLI integration dependencies
typer>=0.9.0
click>=8.1.0
rich>=13.0.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-httpx>=0.26.0
pytest-mock>=3.12.0

# Development and code quality
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.7.0

# Additional utility libraries
python-dotenv>=1.0.0
python-dateutil>=2.8.0
python-jose>=3.3.0
pyyaml>=6.0.0

# Web UI for chat interface
gradio>=4.0.0

# Browser automation for session cookie extraction
playwright>=1.40.0

# Optional: For enhanced async support
trio>=0.23.0
anyio>=4.0.0

# Optional: For production deployment
gunicorn>=21.0.0
prometheus-client>=0.19.0